+++ wiki/CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+## 2026-05-16
+
+- [2.9.1] Bug fix in ```form-select-ext``` widget
+- [2.9.1] Bug fix in ```util.getInputAttr``` helper
+
 ## 2026-05-11
 
 - [2.9.0] Updates to match ```dobo@2.23.0``` specs